+++ st2api/tests/unit/controllers/v1/test_kvps.py
@@ -20,6 +20,11 @@ KVP = {
     'value': 'http://localhost:5000/v3'
 }
 
+KVP_2 = {
+    'name': 'keystone_version',
+    'value': 'v3'
+}
+
 KVP_WITH_TTL = {
     'name': 'keystone_endpoint',
     'value': 'http://localhost:5000/v3',
@@ -40,6 +45,27 @@ class TestKeyValuePairController(FunctionalTest):
         self.assertEqual(get_resp.status_int, 200)
         self.assertEqual(self.__get_kvp_id(get_resp), kvp_id)
         self.__do_delete(kvp_id)
+
+    def test_get_all_prefix_filtering(self):
+        put_resp1 = self.__do_put(KVP['name'], KVP)
+        put_resp2 = self.__do_put(KVP_2['name'], KVP_2)
+        self.assertEqual(put_resp1.status_int, 200)
+        self.assertEqual(put_resp2.status_int, 200)
+
+        # No keys with that prefix
+        resp = self.app.get('/v1/keys?prefix=something')
+        self.assertEqual(resp.json, [])
+
+        # Two keys with the provided prefix
+        resp = self.app.get('/v1/keys?prefix=keystone')
+        self.assertEqual(len(resp.json), 2)
+
+        # One key with the provided prefix
+        resp = self.app.get('/v1/keys?prefix=keystone_endpoint')
+        self.assertEqual(len(resp.json), 1)
+
+        self.__do_delete(self.__get_kvp_id(put_resp1))
+        self.__do_delete(self.__get_kvp_id(put_resp2))
 
     def test_get_one_fail(self):
         resp = self.app.get('/v1/keys/1', expect_errors=True)